V1.6 Latest version running on the emulator. Without errors or warnings.
V1.5
Add black clor background to the actionBar Title.

V1.4
Add new welcomepagestiebeleltronasia.png taken from the Stiebel Eltron Asia web page.

V1.4
Add ic_launcher.png customized icon to the App.

V1.3
In the version 1.3 we have complete the circular menu customization.
We added the functionality and implemented the layouts for:
- WATER FILTER
- HEAT PUMP
- WATER PUMP
- HAND DRYER

V1.2
In the version 1.2 we have customized the circular menu.  The options now has a black color
background to match with the Stiebel Eltron profile.
Also we have customized the:
- Water Heater Page:  adding the background image.
- customized the seekarc to show the progress in GREEN, YELLOW and RED color.
- and finally customized the IL Series with a background image that show the picture
of the IL Series water heater
NEXT STEP: CUSTOMIZE THE OPTIONS TO SELECT:
- WATER FILTER
- HEAT PUMP
- WATER PUMP
- HAND DRYER

V1.1
- add DWHActivity.java
- add activity_dwh.xml
- add selection from floating circular menu for dwh activity
- add display dwh layout
- add seekarc to dwh layout
- the changes were compiled and linked correctly.
- NEXT STEP: CUSTOMIZE DWH LAYOUT WITH STIEBEL ELTRON BACKGROUND.